% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{AIMSmodel}
\alias{AIMSmodel}
\title{The AIMS model}
\format{
An object of class \code{list} of length 4.
}
\usage{
data("AIMSmodel")
}
\value{
\item{all.pairs}{The 100 rules in AIMS in the form EntrezID gene A < EntrezID gene B}
\item{k}{The selected number of optimal rules. For AIMS we have shown it is 20}
\item{one.vs.all.tsp}{The Naive bayes classifier used in combination with the 100 rules}
\item{selected.pairs.list}{The list of rules sorted from the best discriminating rule to the least discriminating rules subdivided by subtype}
}
\description{
This is the model definition for AIMS. It contains the
naive bayes classifier composed of the 100 rules described in Paquet et al.
"Absolute assignment of breast cancer intrinsic molecular subtype" (under
review at JNCI).
}
\details{
This is the AIMS model define using 100 simple rules of the form
gene A < gene B and combine within a naive bayes classifier within e1071.
(Paquet et al. under review JNCI).

Briefly, using a suitably large training dataset(~5000 gene breast cancer
gene expression profiles), the approach identifies a small set of simple
binary rules (~20) that examine the raw expression measurements for pairs of
genes from a single breast cancer patient, and only that patient. The binary
rules are of the form "if the expression of gene x is greater than gene y,
then tend to assign subtype z for that patient". Subtypes could be : Basal,
Her2, LumA, LumB, or Normal. The collection of binary rules is combined for a
single estimation of a patient subtype via a single probabilistic model using
naiveBayes in e1071. In this way, since only expression levels of genes with
a single patient is considered, the method represents a promising approach to
ablate the instability caused by relativistic approaches (Paquet et al. in
review at JNCI).
}
\examples{
library(BreastSubtypeR)
data("AIMSmodel")


}
\references{
\itemize{
\item Paquet ER, Hallett MT. \emph{Absolute assignment of breast cancer
intrinsic molecular subtype.} J Natl Cancer Inst. 2015;107(1).
https://doi.org/10.1093/jnci/dju357
}
}
\keyword{datasets}
