% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_Check}
\alias{BS_Check}
\title{BS_Check}
\usage{
BS_Check(
  data_input,
  phenodata,
  POP = TRUE,
  methods = NA,
  hasClinical = FALSE,
  ...
)
}
\arguments{
\item{data_input}{list of two Gene expression matrix, output of Mapping function; "x_parker" object should be log2 transformed manually.}
}
\value{
The subtypes estimated by selected methods
}
\description{
This calls Ensembl method to predict BC intrinsic subtype.
}
