% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_IHC}
\alias{BS_IHC}
\title{BS_IHC}
\usage{
BS_IHC(gene_expression_matrix, phenodata, hasClinical = FALSE, ...)
}
\arguments{
\item{gene_expression_matrix}{Gene expression matrix, gene in row and sample in column}

\item{phenodata}{Clinical information table, with IHC column}

\item{hasClinical}{Logical. Please specify if you prepared clincical information, like Tumore size as T column, lymphatic node status as NODE column.}

\item{featuredata}{Annotate genes}
}
\value{
The intrinsic subtype, confidential level, percentages.
}
\description{
This predicts breast cancer intrinsic subtypes with ER balanced subset for gene centering.
}
