% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_sspbc}
\alias{BS_sspbc}
\title{Intrinsic Subtyping using SSPBC (BS_sspbc)}
\usage{
BS_sspbc(se_obj, ssp.name = "ssp.pam50")
}
\arguments{
\item{se_obj}{A \code{SummarizedExperiment} object containing:
\itemize{
\item \strong{Assay data}: A gene expression matrix with genes (EntrezID) as rows and samples as columns.
Important: The gene expression values should not be gene-centered.
All expression values must be \strong{positive}.
}}

\item{ssp.name}{Specifies the model to use. Options are:
\itemize{
\item "ssp.pam50": For PAM50-based predictions.
\item "ssp.subtype": For predicting Prosigna-like subtypes (four subtypes, \strong{excluding} the Normal-like subtype).
}}
}
\value{
Returns a vector of intrinsic subtypes assigned to the samples, as estimated
by the SSPBC method.
}
\description{
This function predicts breast cancer intrinsic subtypes using
SSPBC (Single Sample Predictor for Breast Cancer). SSPBC is
based on a refined version of the original AIMS methodology, utilizing a large,
uniformly accrued population-based cohort (SCAN-B) for training. This method supports
RNA sequencing data and provides flexibility in selecting the prediction model.
}
\examples{

# Load required dataset
data("OSLO2EMIT0obj")

# Perform subtyping with the SSPBC method
res <- BS_sspbc(
    se_obj = OSLO2EMIT0obj$data_input$se_SSP,
    ssp.name = "ssp.pam50"
)

}
\references{
\itemize{
\item Staaf J, Häkkinen J, Hegardt C, Saal LH, Kimbung S, Hedenfalk I,
et al. \emph{RNA sequencing-based single sample predictors of molecular subtype and
risk of recurrence for clinical assessment of early-stage breast cancer}. NPJ
Breast Cancer. 2022;8(1). https://doi.org/10.1038/s41523-022-00465-3
}
}
