% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Gene.ID.ann}
\alias{Gene.ID.ann}
\title{Gene annotation table.}
\format{
An object of class \code{data.frame} with 19675 rows and 6 columns.
}
\usage{
data("Gene.ID.ann")
}
\value{
\item{Gene.ID.ann}{Annotation table for GENCODE Human Release 27
genes.}
}
\description{
Annotation table for the GENCODE Human Release 27 genes
(Gene.ID) included in the StringTie target when summarizing gene expression.
The annotation are from from GENCODE Human Release 27 metadata files and
include HGNC, EntrezGene and RefSeq.
}
\details{
Annotation table used by the applySSP to translate gene identifiers
as needed before classification with provided ssp models.
}
\examples{
library(BreastSubtypeR)
## Load the Gene.ID.ann
data(Gene.ID.ann)

}
\references{
\itemize{
\item Staaf J, Häkkinen J, Hegardt C, Saal LH, Kimbung S, Hedenfalk I,
et al. \emph{RNA sequencing-based single sample predictors of molecular subtype and
risk of recurrence for clinical assessment of early-stage breast cancer}. NPJ
Breast Cancer. 2022;8(1). https://doi.org/10.1038/s41523-022-00465-3
}
}
\keyword{datasets}
