% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneID_mapping.R
\name{GeneID_mapping}
\alias{GeneID_mapping}
\title{Functions for mapping Gene ID}
\usage{
GeneID_mapping(test)
}
\arguments{
\item{name}{description}
}
\value{
The subtypes estimated by genefu robust
}
\description{
Functions for mapping Gene ID
}
\examples{
# example code


}
