% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{get_average_subtype}
\alias{get_average_subtype}
\title{Function to get the average correlation and ROR}
\usage{
get_average_subtype(res_ihc_iterative, consensus_subtypes)
}
\description{
Function to get the average correlation and ROR
}
