% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{OSLO2EMIT0obj}
\alias{OSLO2EMIT0obj}
\title{OSLO2EMIT0obj: Example Dataset for the OSLO2EMIT0 Cohort}
\format{
A list containing the following elements:
\describe{
\item{\code{se_obj}}{A SummarizedExperiment object containing a subset of
the gene expression matrix, clinical information, and gene feature
annotations for the OSLO2EMIT0 cohort.}
\item{\code{data_input}}{Example output from the \code{Mapping} function.}
\item{\code{res}}{Example output from the \code{BS_Multi} function.}
}
}
\usage{
data("OSLO2EMIT0obj")
}
\description{
This example dataset is based on the OSLO2EMIT0 cohort
described in Staaf et al., 2022. It contains subsetted data for gene
expression, clinical information, feature annotations, and example outputs
from the \code{Mapping} and \code{BS_Multi} functions.
}
\examples{
library(BreastSubtypeR)
data("OSLO2EMIT0obj")

}
\references{
\itemize{
\item Staaf J, Häkkinen J, Hegardt C, Saal LH, Kimbung S, Hedenfalk I,
et al. \emph{RNA sequencing-based single sample predictors of molecular subtype and
risk of recurrence for clinical assessment of early-stage breast cancer}. NPJ
Breast Cancer. 2022;8(1). https://doi.org/10.1038/s41523-022-00465-3
}
}
\keyword{datasets}
