% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCIS_methods.R
\name{PAM50_AIMS}
\alias{PAM50_AIMS}
\title{PAM50_AIMS}
\usage{
PAM50_AIMS(gene_expression_matrix, EntrezID, ...)
}
\arguments{
\item{gene_expression_matrix}{Gene expression matrix, gene in row and sample in column}

\item{featuredata}{Annotate genes}

\item{phenodata}{Clinical information table, with ER information or IHC column}
}
\value{
The subtypes estimated by PCA_AIMS
}
\description{
This calls AIMS to do PAM50 subtyping.
}
