% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM50_subtyping_methods.R
\name{PAM50_ensembl}
\alias{PAM50_ensembl}
\alias{PAM50_Ensembl}
\title{PAM50_ensembl}
\usage{
PAM50_Ensembl(
  data_input,
  phenodata,
  POP = TRUE,
  methods = NA,
  hasClinical = FALSE,
  ...
)
}
\arguments{
\item{data_input}{list of two Gene expression matrix, output of Mapping function}
}
\value{
The subtypes estimated by selected methods
}
\description{
This calls Ensembl method to predict BC intrinsic subtype.
}
