% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM50_subtyping_methods.R
\name{PAM50_IHC.itr}
\alias{PAM50_IHC.itr}
\title{PAM50_IHC.itr}
\usage{
PAM50_IHC.itr(
  gene_expression_matrix,
  phenodata,
  iterative = 100,
  ratio = 54/64,
  hasClinical = FALSE,
  ...
)
}
\arguments{
\item{gene_expression_matrix}{Gene expression matrix, gene in row and sample in column}

\item{phenodata}{Clinical information table, with IHC column}

\item{iterative}{times to do iterative ER balanced procedure with certain ratio}

\item{ratio}{The options are either 1:1 or 54(ER+):64(ER-) (change it as default). The latter is ER ratio used for UNC230 train cohort}

\item{hasClinical}{Logical. Please specify if you prepared clincical information, like Tumore size as T column, lymphatic node status as NODE column.}

\item{featuredata}{Annotate genes}
}
\value{
The intrinsic subtype, confidential level, percentages.
}
\description{
This predicts intrinsic breast cancer subtypes with ER balanced subset for gene centering
}
