% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM50_subtyping_methods.R
\name{PAM50_IHC}
\alias{PAM50_IHC}
\title{PAM50_IHC}
\usage{
PAM50_IHC(gene_expression_matrix, phenodata, hasClinical = FALSE, ...)
}
\arguments{
\item{gene_expression_matrix}{Gene expression matrix, gene in row and sample in column}

\item{phenodata}{Clinical information table, with IHC column}

\item{hasClinical}{Logical. Please specify if you prepared clincical information, like Tumore size as T column, lymphatic node status as NODE column.}

\item{featuredata}{Annotate genes}
}
\value{
The intrinsic subtype, confidential level, percentages.
}
\description{
This predicts intrinsic breast cancer subtypes with ER balanced subset for gene centering
}
