% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCIS_methods.R
\name{PAM50_parker}
\alias{PAM50_parker}
\title{PAM50_parker}
\usage{
PAM50_parker(
  gene_expression_matrix,
  phenodata,
  calibration = "None",
  internal = NA,
  external = NA,
  medians = NA,
  hasClinical = FALSE,
  ...
)
}
\arguments{
\item{gene_expression_matrix}{Gene expression matrix, gene in row and sample in column, log of normalized}

\item{phenodata}{Clinical information table, with ER information or IHC column (what input)}

\item{calibration}{How to do calibration, "None"(default) means no calibration for gene expression matrix. When setting calibration =None, you dont need to set internal and external parameters.  "Internal" means calibration for gene expression matrix by itself. "External" means calibration by external cohort.}

\item{internal}{Specify the strategy for internal calibration, medianCtr(default), meanCtr and qCtr}

\item{external}{Specify the platform name(which column) of external medians calculated by train cohorts. When users want to use Medians prepared by user selves, this parameter should be "Given.mdns", not platform name.}

\item{medians}{If you specify "external" parameter as "Given.mdns", you should input matrix/table, 50 signatures in the first column and "Given.mdns" values in the second column.}

\item{hasClinical}{Logical. Please specify if you prepared clinical information, like Tumore size as T column, lymphatic node status as NODE column.}

\item{featuredata}{Annotate genes}

\item{do.mapping}{Logical. If it is microarray data or expression matrix at transcript level, please specify TRUE (this will call dochecking() function and run mapping step )}
}
\value{
The subtypes estimated by parker-based PAM50 subtyping
}
\description{
This calls parker-based PAM50 subtyping methods, including None, medianCtr, meanCtr, qCtr, Given.mdns, or chosen platform to do calibration.
}
