% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM50_subtyping_methods.R
\name{PAM50_PCA-PAM50}
\alias{PAM50_PCA-PAM50}
\alias{PAM50_PCA_PAM50}
\title{PAM50_PCA}
\usage{
PAM50_PCA_PAM50(gene_expression_matrix, phenodata, hasClinical = FALSE)
}
\arguments{
\item{gene_expression_matrix}{Gene expression matrix, gene in row and sample in column}

\item{phenodata}{Clinical information table, with ER information or IHC column}

\item{hasClinical}{Logical. Please specify if you prepared clincical information, like Tumore size as T column, lymphatic node status as NODE column.}

\item{featuredata}{Annotate genes}
}
\value{
The intrinsic subtypes estimated by PCA-PAM50
}
\description{
This calls PCA-PAM50 (Raj-Kumar, PK.) to do PAM50 subtyping.
}
