% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCIS_methods.R
\name{PAM50_ssBC}
\alias{PAM50_ssBC}
\title{PAM50_ssBC}
\usage{
PAM50_ssBC(gene_expression_matrix, phenodata, s, hasClinical = FALSE, ...)
}
\arguments{
\item{gene_expression_matrix}{Gene expression matrix, gene in row and sample in column}

\item{phenodata}{Clinical information table, like TN,ER and HER2}

\item{s}{Specify "ER" or "TN", which is included in phenodata}

\item{hasClinical}{Logical. Please specify if you prepared clincical information, like Tumore size as T column, lymphatic node status as NODE column.}

\item{featuredata}{Annotate genes}
}
\value{
The subtypes estimated by ssBC
}
\description{
This calls ssBC to do PAM50 subtyping.
It is better to group cohorts by ER+/- or TN/nTN.
}
