% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM50_subtyping_methods.R
\name{PAM50_sspbc}
\alias{PAM50_sspbc}
\title{PAM50_sspbc}
\usage{
PAM50_sspbc(gene_expression_matrix, ...)
}
\arguments{
\item{gene_expression_matrix}{Gene expression matrix, gene in row sample in column}
}
\value{
The subtypes estimated by sspbc
}
\description{
This calls sspbc to do PAM50 subtyping.
}
