% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BCIS_methods.R
\docType{package}
\name{PAM50subtyping}
\alias{PAM50subtyping}
\title{Collection of breast cancer PAM50 subtyping methods}
\description{
PAM50subtyping is a R package that integrates the access to PAM50 subtyping methods and
visualization ...
}
\details{
Collection of breast cancer PAM50 subtyping methods
}
