% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{prepare_ssp_matrix}
\alias{prepare_ssp_matrix}
\title{Prepare matrix for SSP-based methods}
\usage{
prepare_ssp_matrix(x, genes.s, RawCounts, samplenames, verbose)
}
\arguments{
\item{x}{Numeric matrix (rows = Entrez IDs, cols = samples).}

\item{genes.s}{Dataframe with gene signatures.}

\item{RawCounts}{Logical; if TRUE, input is raw counts.}

\item{verbose}{Logical; print missing genes.}
}
\value{
Processed matrix for SSP.
}
\description{
Prepare matrix for SSP-based methods
}
