% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyBreastSubtypeR.R
\name{runShinyBreastSubtypeR}
\alias{runShinyBreastSubtypeR}
\title{(Deprecated) Run iBreastSubtypeR}
\usage{
runShinyBreastSubtypeR(...)
}
\arguments{
\item{...}{Arguments passed on to \code{\link[=iBreastSubtypeR]{iBreastSubtypeR()}}.}
}
\description{
Back-compat wrapper; use \code{\link[=iBreastSubtypeR]{iBreastSubtypeR()}} instead.
}
