% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{sspbc.models.fullname}
\alias{sspbc.models.fullname}
\title{sspbc.models.fullname: Full names for 11 SSPBC predictors}
\format{
An object of class \code{list} of length 11.
}
\usage{
data("sspbc.models.fullname")
}
\value{
\item{sspbc.models.fullname}{Named list of 11 SSP models used by \code{sspbc}.}
}
\description{
List of the same 11 SSP models (Staaf et al., 2022) indexed by full model names.
}
\details{
Identical content to \code{sspbc.models} but with full model names as list keys.
}
\examples{
library(BreastSubtypeR)
data("sspbc.models.fullname")
}
\references{
Staaf J, Häkkinen J, Hegardt C, Saal LH, Kimbung S, Hedenfalk I, et al.
RNA sequencing-based single sample predictors of molecular subtype and risk of
recurrence for clinical assessment of early-stage breast cancer.
\emph{NPJ Breast Cancer}. 2022;8(1):27. https://doi.org/10.1038/s41523-022-00465-3
}
\keyword{datasets}
