% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC_functions.R
\name{subtypePrediction_distributed}
\alias{subtypePrediction_distributed}
\title{Function for predicting PAM50 intrinsic subtypes and calculation of proliferation}
\usage{
subtypePrediction_distributed(
  mat,
  df.al,
  calibration = "None",
  internal = NA,
  external = NA,
  short = "PAM50",
  hasClinical = FALSE,
  vis = FALSE
)
}
\description{
Function for predicting PAM50 intrinsic subtypes and calculation of proliferation
}
