% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Vis_boxpot}
\alias{Vis_boxpot}
\title{Functions for visualization
Function for boxplot of correlation per subtype}
\usage{
Vis_boxpot(out, correlations)
}
\arguments{
\item{out}{a data frame includes "patientID" and "Subtype"}

\item{correlations}{correlations table from NC-based methods}
}
\description{
Functions for visualization
Function for boxplot of correlation per subtype
}
\examples{


data("OSLO2MEITOobj")

out= data.frame(PatientID = res$results$parker.original$BS.all$PatientID, Subtype = res$results$parker.original$BS.all$BS )
correlations =res$results$parker.original$outList$distances

p = Vis_boxpot(out, correlations )
plot(p)

}
