% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Vis_consensus}
\alias{Vis_consensus}
\title{Function for pieplot}
\usage{
Vis_consensus(data)
}
\arguments{
\item{out}{a data table includes "patientID" and "Subtype"}
}
\description{
Function for pieplot
}
