% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Vis_Multi}
\alias{Vis_Multi}
\title{Multi-Method Subtype Heatmap Visualization}
\usage{
Vis_Multi(data)
}
\arguments{
\item{data}{Output of the \code{\link{BS_Multi}} function.}
}
\value{
Returns a heatmap visualizing the subtype classifications across multiple methods.
}
\description{
This function generates a heatmap to visualize breast cancer subtypes
classified by multiple subtyping methods. It helps users compare how different methods
assign subtypes to the same set of samples.
}
\examples{
data("OSLO2EMIT0obj")

# Assuming `OSLO2EMIT0obj$res$res_subtypes` contains multi-method subtype results
p <- Vis_Multi(OSLO2EMIT0obj$res$res_subtypes)
plot(p)

}
