% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Vis_PCA}
\alias{Vis_PCA}
\title{Function for PCAplot}
\usage{
Vis_PCA(x, out, Eigen = FALSE)
}
\arguments{
\item{x}{gene expression matrix, log2 transformed}

\item{out}{a data table includes "patientID" and "Subtype"}

\item{screeplot}{Logic. Please specify if show screeplot}
}
\description{
Function for PCAplot
}
