% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Vis_pie}
\alias{Vis_pie}
\title{Function for pieplot}
\usage{
Vis_pie(out)
}
\arguments{
\item{out}{a data table includes "patientID" and "Subtype"}
}
\description{
Function for pieplot
}
\examples{

data("OSLO2MEITOobj")

out= data.frame(PatientID = rownames(res$results$sspbc$BS.all), Subtype = res$results$sspbc$BS.all$BS )
plot = Vis_pie( out = out)

}
