% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\docType{package}
\name{BreastSubtypeR}
\alias{BreastSubtypeR}
\title{Collection of breast cancer intrinsic subtyping methods}
\description{
BreastSubtypeR is an R package that integrates the access to intrinsic subtyping methods.
}
\details{
Collection of breast cancer intrinsic subtyping methods
}
