% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_AIMS}
\alias{BS_AIMS}
\title{BS_AIMS}
\usage{
BS_AIMS(gene_expression_matrix, EntrezID, ...)
}
\arguments{
\item{gene_expression_matrix}{A gene expression matrix with genes in rows and samples in columns. The data should be log-transformed.}

\item{EntrezID}{A list of Entrez gene ids}
}
\value{
The subtypes estimated by AIMS method
}
\description{
This calls AIMS to do intrinsic subtyping.
}
\examples{

data("BreastSubtypeR")
data("OSLO2MEITOobj")

genes = as.character( BreastSubtypeR$genes.signature$EntrezGene.ID[which( BreastSubtypeR$genes.signature$AIMS == "Yes" )])
res = BS_AIMS(data_input$x_SSP[genes,], rownames(data_input$x_SSP[genes,]) )

}
\references{
Paquet ER, Hallett MT. Absolute assignment of breast cancer intrinsic molecular subtype. J Natl Cancer Inst. 2015;107(1). https://doi.org/10.1093/jnci/dju357
}
