% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_AIMS}
\alias{BS_AIMS}
\title{AIMS Intrinsic Subtyping (BS_AIMS)}
\usage{
BS_AIMS(se_obj)
}
\arguments{
\item{se_obj}{A \code{SummarizedExperiment} object containing:
\itemize{
\item \strong{Assay data}: A gene expression matrix with genes (EntrezID) as rows and samples as columns.
Important: The gene expression values should not be gene-centered.
All expression values must be \strong{positive}.
}}
}
\value{
Returns a vector of intrinsic subtypes assigned to the samples, as estimated
by the AIMS method.
}
\description{
This function predicts breast cancer intrinsic subtypes using
the \strong{AIMS (Absolute assignment of Intrinsic Molecular Subtype)} method.
}
\examples{
# Load required datasets
data("OSLO2EMIT0obj")


# Perform subtyping
res <- BS_AIMS(
    se_obj = OSLO2EMIT0obj$data_input$se_SSP
)

}
\references{
\itemize{
\item Paquet ER, Hallett MT. \emph{Absolute assignment of breast cancer
intrinsic molecular subtype.} J Natl Cancer Inst. 2015;107(1).
https://doi.org/10.1093/jnci/dju357
}
}
