% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_Check}
\alias{BS_Check}
\title{BS_Check}
\usage{
BS_Check(
  data_input,
  phenodata,
  methods = NA,
  Prosigna = FALSE,
  hasClinical = FALSE,
  ...
)
}
\arguments{
\item{phenodata}{A clinical information table. The first column must be named "PatientID".}

\item{methods}{Specify methods.}

\item{Prosigna}{Logic.}

\item{hasClinical}{Logical. Specify whether clinical information is included. For example, tumor size should be in the "T" column, and lymph node status should be in the "NODE" column.}

\item{gene_expression_matrix}{A gene expression matrix with genes in rows and samples in columns. The data should be log-transformed.}
}
\value{
The subtypes estimated by selected methods
}
\description{
This calls consensus subtyping and users can specify subtyping methods.
}
\examples{

methods = c( "parker.median", "parker.mean", "parker.quantile")
res = BS_Check(data_input = data_input, phenodata = clinic.oslo, methods = methods, Prosigna = FALSE, hasClinical = FALSE)


}
