% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_cIHC.itr}
\alias{BS_cIHC.itr}
\title{BS_cIHC.itr}
\usage{
BS_cIHC.itr(
  gene_expression_matrix,
  phenodata,
  iteration = 100,
  ratio = 54/64,
  Prosigna = FALSE,
  hasClinical = FALSE,
  seed = 118,
  ...
)
}
\arguments{
\item{gene_expression_matrix}{A gene expression matrix with genes in rows and samples in columns. The data should be log-transformed.}

\item{phenodata}{A clinical information table. The first column must be named "PatientID".}

\item{ratio}{The options are either 1:1 or 54 (ER+) : 64 (ER-) (default). The latter was ER ratio used for UNC230 train cohort.}

\item{Prosigna}{Logic.}

\item{hasClinical}{Logical. Specify whether clinical information is included. For example, tumor size should be in the "T" column, and lymph node status should be in the "NODE" column.}

\item{iterative}{Times to do iterative ER balanced procedure with certain ratio.}
}
\value{
The intrinsic subtypes, confidential level, percentages.
}
\description{
This predicts breast cancer intrinsic subtypes with ER subset for gene centering. It supports the selection of ER subset ratio.
}
\examples{

data("oslo.obj")
res =  BS_cIHC.itr(data_input$x_NC.log, phenodata = clinic.oslo,  Prosigna = FALSE, hasClinical =FALSE)

}
