% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_cIHC.itr}
\alias{BS_cIHC.itr}
\title{BS_cIHC.itr}
\usage{
BS_cIHC.itr(
  gene_expression_matrix,
  phenodata,
  iteration = 100,
  ratio = 54/64,
  Subtype = FALSE,
  hasClinical = FALSE,
  seed = 118,
  ...
)
}
\arguments{
\item{gene_expression_matrix}{A gene expression matrix with genes in rows and samples in columns. The data should be log-transformed.}

\item{phenodata}{A clinical information table. The first column must be named "PatientID".}

\item{ratio}{The options are either 1:1 or 54 (ER+) : 64 (ER-) (default). The latter was ER ratio used for UNC230 train cohort.}

\item{Subtype}{Logic. Specify whether to predict four subtypes by removing the Normal-like subtype.}

\item{hasClinical}{Logic. Specify whether clinical information is included. For example, tumor size should be in the "T" column, and lymph node status should be in the "NODE" column.}

\item{iterative}{Times to do iterative ER balanced procedure with certain ratio.}
}
\value{
The intrinsic subtypes, confidential level, percentages.
}
\description{
This predicts breast cancer intrinsic subtypes with ER subset for gene centering. It supports the selection of ER subset ratio.
}
\examples{

data("OSLO2MEITOobj")
res =  BS_cIHC.itr(data_input$x_NC.log, phenodata = clinic.oslo,  Subtype = FALSE, hasClinical =FALSE)

}
\references{
Curtis C, Shah SP, Chin SF, Turashvili G, Rueda OM, Dunning MJ, et al. The genomic and transcriptomic architecture of 2,000 breast tumours reveals novel subgroups. Nature. 2012;486(7403). https://doi.org/10.1038/nature10983
}
