% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_cIHC.itr}
\alias{BS_cIHC.itr}
\title{Iterative conventional IHC Intrinsic Subtyping (BS_cIHC.itr)}
\usage{
BS_cIHC.itr(
  se_obj,
  iteration = 100,
  ratio = 54/64,
  Subtype = FALSE,
  hasClinical = FALSE,
  seed = 118
)
}
\arguments{
\item{se_obj}{A \code{SummarizedExperiment} object containing:
\itemize{
\item \strong{Assay data}: A gene expression matrix with genes (Gene Symbol)
in rows and samples in columns. This matrix should be log-transformed.
\item \strong{Column metadata} (\code{colData}): Clinical information table.
\itemize{
\item \code{"PatientID"}: Unique sample or patient identifiers.
\item \code{"ER"}: Estrogen receptor (ER) status recorded as \code{"ER+"} or \code{"ER-"}.
}
}}

\item{iteration}{Integer. Number of iterations for the ER-balanced procedure
with the specified ratio. The default is 100.}

\item{ratio}{Numeric. Specifies the ER+/ER??? ratio for balancing. Options
are \code{1:1} or \code{54:64} (default). The latter corresponds to the ER+/ER- ratio
used in the UNC232 training cohort.}

\item{Subtype}{Logical. If \code{TRUE}, the function predicts four subtypes by
excluding the Normal-like subtype.}

\item{hasClinical}{Logical. If \code{TRUE}, the function uses clinical data from
the \code{pheno} table. Required columns include:
\itemize{
\item \code{"TSIZE"}: Tumor size.
\item \code{"NODE"}: Lymph node status.
}}

\item{seed}{An integer value is used to set the random seed.}
}
\value{
A list containing:
\itemize{
\item Intrinsic subtype predictions.
\item Confidence levels for each subtype.
\item Percentages of ER+ and ER??? subsets across iterations.
}
}
\description{
This function predicts breast cancer intrinsic subtypes using an
iterative ER-balanced procedure for gene centering. It supports customization
of the ER+/ER- ratio.
}
\examples{
data("OSLO2EMIT0obj")
res <- BS_cIHC.itr(
    se_obj = OSLO2EMIT0obj$data_input$se_NC,
    iteration = 10, ## For final analysis, set iteration = 100
    Subtype = FALSE,
    hasClinical = FALSE
)

}
\references{
\itemize{
\item Curtis C, Shah SP, Chin SF, Turashvili G, Rueda OM, Dunning MJ,
et al. \emph{The genomic and transcriptomic architecture of 2,000 breast tumours
reveals novel subgroups}. Nature. 2012;486(7403).
https://doi.org/10.1038/nature10983
}
}
