% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_cIHC.itr}
\alias{BS_cIHC.itr}
\title{Iterative conventional IHC Intrinsic Subtyping (BS_cIHC.itr)}
\usage{
BS_cIHC.itr(
  gene_expression_matrix,
  phenodata,
  iteration = 100,
  ratio = 54/64,
  Subtype = FALSE,
  hasClinical = FALSE,
  seed = 118,
  ...
)
}
\arguments{
\item{gene_expression_matrix}{A gene expression matrix with genes in rows and samples in columns. The data should be log-transformed.}

\item{phenodata}{A clinical information table. The first column must contain sample or patient names and be named \code{"PatientID"}. Additionally, the table must include an \code{"ER"} column, where estrogen receptor (ER) status is recorded as \code{"ER+"} or \code{"ER-"}.}

\item{ratio}{Numeric. Specifies the ER+ to ER− ratio for balancing. Options are \code{1:1} or \code{54:64} (default). The latter corresponds to the ER ratio used in the UNC230 training cohort.}

\item{Subtype}{Logical. If \code{TRUE}, the function predicts four subtypes by excluding the Normal-like subtype.}

\item{hasClinical}{Logical. If \code{TRUE}, the function uses clinical data from the \code{phenodata} table. Required columns include:
\itemize{
\item \code{"T"}: Tumor size.
\item \code{"NODE"}: Lymph node status.
}}

\item{iterative}{Integer. Number of iterations for the ER-balanced procedure with a specified ratio.}
}
\value{
A list containing:
\itemize{
\item Intrinsic subtype predictions.
\item Confidence levels for each subtype.
\item Percentages of ER+ and ER− subsets across iterations.
}
}
\description{
Predicts breast cancer intrinsic subtypes using an iterative ER-balanced procedure for gene centering. Supports customization of ER subset ratios.
}
\examples{
data("OSLO2EMITOobj")
res <- BS_cIHC.itr(
  gene_expression_matrix = data_input$x_NC.log,
  phenodata = clinic.oslo,
  Subtype = FALSE,
  hasClinical = FALSE
)

}
\references{
Curtis C, Shah SP, Chin SF, Turashvili G, Rueda OM, Dunning MJ, et al. \emph{The genomic and transcriptomic architecture of 2,000 breast tumours reveals novel subgroups}. Nature. 2012;486(7403). https://doi.org/10.1038/nature10983
}
