% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_cIHC}
\alias{BS_cIHC}
\title{Conventional IHC Intrinsic Subtyping (BS_cIHC)}
\usage{
BS_cIHC(se_obj, Subtype = FALSE, hasClinical = FALSE, seed = 118)
}
\arguments{
\item{se_obj}{A \code{SummarizedExperiment} object containing:
\itemize{
\item \strong{Assay data}: A gene expression matrix with genes (Gene Symbol)
in rows and samples in columns. This matrix should be log-transformed.
\item \strong{Column metadata} (\code{colData}): Clinical information table.
\itemize{
\item \code{"PatientID"}: Unique sample or patient identifiers.
\item \code{"ER"}: Estrogen receptor (ER) status recorded as \code{"ER+"} or \code{"ER-"}.
}
}}

\item{Subtype}{Logical. If \code{TRUE}, the function predicts four subtypes by
excluding the Normal-like subtype.}

\item{hasClinical}{Logical. If \code{TRUE}, the function uses clinical data from
the \code{pheno} table. Required columns include:
\itemize{
\item \code{"TSIZE"}: Tumor size.
\item \code{"NODE"}: Lymph node status.
}}

\item{seed}{An integer value is used to set the random seed.}
}
\value{
A data frame containing the intrinsic subtypes estimated using the
conventional IHC (cIHC) method.
}
\description{
This function predicts breast cancer intrinsic subtypes using
the conventional IHC (cIHC) method, which employs ER-balanced subsets for
gene centering.
}
\examples{
data("OSLO2EMIT0obj")
res <- BS_cIHC(
    se_obj = OSLO2EMIT0obj$data_input$se_NC,
    Subtype = FALSE,
    hasClinical = FALSE
)

}
\references{
\itemize{
\item Ciriello G, Gatza ML, Beck AH, Wilkerson MD, Rhie SK, Pastore A,
et al. \emph{Comprehensive Molecular Portraits of Invasive Lobular Breast Cancer}.
Cell. 2015;163(2). https://doi.org/10.1016/j.cell.2015.09.033
}
}
