% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_cIHC}
\alias{BS_cIHC}
\title{BS_cIHC}
\usage{
BS_cIHC(
  gene_expression_matrix,
  phenodata,
  Subtype = FALSE,
  hasClinical = FALSE,
  ...
)
}
\arguments{
\item{gene_expression_matrix}{A gene expression matrix with genes in rows and samples in columns. The data should be log-transformed.}

\item{phenodata}{A clinical information table. The first column must be named "PatientID".}

\item{Subtype}{Logic. Specify whether to predict four subtypes by removing the Normal-like subtype.}

\item{hasClinical}{Logic. Specify whether clinical information is included. For example, tumor size should be in the "T" column, and lymph node status should be in the "NODE" column.}
}
\value{
The intrinsic subtypes estimated using the conventinal IHC (cIHC) method.
}
\description{
This predicts breast cancer intrinsic subtypes with ER balanced subset for gene centering.
}
\examples{

data("OSLO2MEITOobj")
res = BS_cIHC(data_input$x_NC.log, phenodata= clinic.oslo,  Subtype = FALSE, hasClinical =FALSE)

}
\references{
Ciriello G, Gatza ML, Beck AH, Wilkerson MD, Rhie SK, Pastore A, et al. Comprehensive Molecular Portraits of Invasive Lobular Breast Cancer. Cell. 2015;163(2). https://doi.org/10.1016/j.cell.2015.09.033
}
