% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_cIHC}
\alias{BS_cIHC}
\title{Conventional IHC Intrinsic Subtyping (BS_cIHC)}
\usage{
BS_cIHC(se_obj, Subtype = FALSE, hasClinical = FALSE, seed = 118)
}
\arguments{
\item{se_obj}{A \code{SummarizedExperiment} object containing:
\itemize{
\item \strong{Assay data}: A log-transformed, normalized gene expression matrix with genes (Gene Symbols)
as rows and samples as columns.
\item \strong{Column metadata} (\code{colData}): A clinical information table, which must include:
\itemize{
\item \code{"PatientID"}: Unique sample or patient identifiers.
\item \code{"ER"}: Estrogen receptor (ER) status, recorded as \code{"ER+"} or \code{"ER-"}.
}
}}

\item{Subtype}{Logical (\code{TRUE} or \code{FALSE}). If \code{TRUE}, the function predicts four subtypes,
\strong{excluding} the Normal-like subtype.}

\item{hasClinical}{Logical (\code{TRUE} or \code{FALSE}). If \code{TRUE}, the function incorporates clinical data from
the phenotype (\code{pheno}) table. Required columns:
\itemize{
\item \code{"TSIZE"}: Tumor size (\code{0} for <= 2cm, \code{1} for > 2cm).
\item \code{"NODE"}: Lymph node status (\code{0} for negative, \code{1} or higher for positive nodes; this column must be numeric).
}}

\item{seed}{An integer used to set the random seed for reproducibility.}
}
\value{
Returns a data frame containing intrinsic subtypes estimated using the
conventional IHC (cIHC) method.
}
\description{
This function predicts breast cancer intrinsic subtypes using
the conventional estrogen receptor (ER)-balancing via immunohistochemistry (cIHC).
}
\examples{
data("OSLO2EMIT0obj")
res <- BS_cIHC(
    se_obj = OSLO2EMIT0obj$data_input$se_NC,
    Subtype = FALSE,
    hasClinical = FALSE
)

}
\references{
\itemize{
\item Ciriello G, Gatza ML, Beck AH, Wilkerson MD, Rhie SK, Pastore A,
et al. \emph{Comprehensive Molecular Portraits of Invasive Lobular Breast Cancer}.
Cell. 2015;163(2). https://doi.org/10.1016/j.cell.2015.09.033
}
}
