% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_cIHC}
\alias{BS_cIHC}
\title{BS_cIHC}
\usage{
BS_cIHC(
  gene_expression_matrix,
  phenodata,
  Prosigna = FALSE,
  hasClinical = FALSE,
  ...
)
}
\arguments{
\item{gene_expression_matrix}{A gene expression matrix with genes in rows and samples in columns. The data should be log-transformed.}

\item{phenodata}{A clinical information table. The first column must be named "PatientID".}

\item{Prosigna}{Logic.}

\item{hasClinical}{Logical. Specify whether clinical information is included. For example, tumor size should be in the "T" column, and lymph node status should be in the "NODE" column.}
}
\value{
The intrinsic subtypes estimated using the conventinal IHC (cIHC) method.
}
\description{
This predicts breast cancer intrinsic subtypes with ER balanced subset for gene centering.
}
\examples{

data("oslo.obj")
res = BS_cIHC(data_input$x_NC.log, phenodata= clinic.oslo,  Prosigna = FALSE, hasClinical =FALSE)

}
