% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_IHC.itr}
\alias{BS_IHC.itr}
\title{BS_IHC.itr}
\usage{
BS_IHC.itr(
  gene_expression_matrix,
  phenodata,
  iterative = 100,
  ratio = 54/64,
  hasClinical = FALSE,
  ...
)
}
\arguments{
\item{gene_expression_matrix}{Gene expression matrix, gene in row and sample in column}

\item{phenodata}{Clinical information table, with IHC column}

\item{iterative}{Times to do iterative ER balanced procedure with certain ratio}

\item{ratio}{The options are either 1:1 or 54(ER+):64(ER-) (default). The latter wass ER ratio used for UNC230 train cohort.}

\item{hasClinical}{Logical. Please specify if you prepared clincical information, like Tumore size as "T" column, lymphatic node status as "NODE" column.}

\item{featuredata}{Annotate genes}
}
\value{
The intrinsic subtypes, confidential level, percentages.
}
\description{
This predicts breast cancer intrinsic subtypes with ER subset for gene centering. It supports the selection of ER subset ratio.
}
