% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_Multi}
\alias{BS_Multi}
\title{Intrinsic Subtyping with Multiple Approaches (BS_Multi)}
\usage{
BS_Multi(data_input, methods = "AUTO", Subtype = FALSE, hasClinical = FALSE)
}
\arguments{
\item{data_input}{The output from the \code{Mapping()} function, containing
processed gene expression data prepared for subtyping analysis.}

\item{methods}{A character vector specifying the subtyping methods to be
used. Available options:
\itemize{
\item "parker.original": Original PAM50 by Parker et al., 2009 (Parker et al., 2009)
\item "genefu.scale": PAM50 implementation as in the genefu R package (scaled version) (Gendoo et al., 2016)
\item "genefu.robust": PAM50 implementation as in the genefu R package (robust version) (Gendoo et al., 2016)
\item "cIHC": Conventional estrogen receptor (ER)-balancing using immunohistochemistry (cIHC) (Ciriello et al., 2015)
\item "cIHC.itr": Iterative version of cIHC (Curtis et al., 2012)
\item "PCAPAM50": PCA-based iterative PAM50 (ER-balancing using ESR1 gene expression) (Raj-Kumar et al., 2019)
\item "ssBC": Subgroup-specific gene-centering PAM50 (Zhao et al., 2015)
\item "ssBC.v2": Updated subgroup-specific gene-centering PAM50 with refined quantiles (Fernandez-Martinez et al., 2020)
\item "AIMS": Absolute Intrinsic Molecular Subtyping (AIMS) method (Paquet & Hallett, 2015)
\item "sspbc": Single-Sample Predictors for Breast Cancer (AIMS adaptation) (Staaf et al., 2022)
\item "AUTO": Automatically selects subtyping methods based on the ER/HER2 distribution of the test cohort.
}

Notes:
\itemize{
\item If "AUTO" is selected, it must be the sole value in the vector.
\item If "AUTO" is not selected, at least \strong{two} methods must be specified; otherwise, an error will occur.
}}

\item{Subtype}{Logical (\code{TRUE} or \code{FALSE}). If \code{TRUE}, the function predicts four subtypes,
\strong{excluding} the Normal-like subtype.}

\item{hasClinical}{Logical (\code{TRUE} or \code{FALSE}). If \code{TRUE}, the function incorporates clinical data from
the phenotype (\code{pheno}) table. Required columns:
\itemize{
\item \code{"TSIZE"}: Tumor size (\code{0} for <= 2cm, \code{1} for > 2cm).
\item \code{"NODE"}: Lymph node status (\code{0} for negative, \code{1} or higher for positive nodes; this column must be numeric).
}}
}
\value{
Returns a list of intrinsic subtypes estimated by the selected methods.
}
\description{
This function predicts breast cancer intrinsic subtypes using multiple methods.
Users can either specify the subtyping approaches directly or enable automatic selection ("AUTO")
based on the ER/HER2 distribution of the test cohort.
}
\examples{
# Load required dataset
data("OSLO2EMIT0obj")

# Define methods to use for consensus subtyping
methods <- c("parker.original", "genefu.scale", "genefu.robust")

# Perform subtyping
res.test <- BS_Multi(
    data_input = OSLO2EMIT0obj$data_input,
    methods = methods,
    Subtype = FALSE,
    hasClinical = FALSE
)

}
\references{
\itemize{
\item Parker JS, Mullins M, Cheung MCU, Leung S, Voduc D, et al.
\emph{Supervised risk predictor of breast cancer based on intrinsic subtypes}.
Journal of Clinical Oncology. 2009;27(8).
https://doi.org/10.1200/JCO.2008.18.1370
\item Gendoo DMA, Ratanasirigulchai N, Schröder MS, Paré L, Parker JS, Prat A, et
al. \emph{Genefu: An R/Bioconductor package for computation of gene
expression-based signatures in breast cancer}. Bioinformatics. 2016;32(7).
https://doi.org/10.1093/bioinformatics/btv693
\item Ciriello G, Gatza ML, Beck AH, Wilkerson MD, Rhie SK, Pastore A,
et al. \emph{Comprehensive Molecular Portraits of Invasive Lobular Breast Cancer}.
Cell. 2015;163(2). https://doi.org/10.1016/j.cell.2015.09.033
\item Curtis C, Shah SP, Chin SF, Turashvili G, Rueda OM, Dunning MJ,
et al. \emph{The genomic and transcriptomic architecture of 2,000 breast tumours
reveals novel subgroups}. Nature. 2012;486(7403).
https://doi.org/10.1038/nature10983
\item Zhao X, Rodland EA, Tibshirani R, Plevritis S. \emph{Molecular
subtyping for clinically defined breast cancer subgroups.} Breast Cancer
Research. 2015;17(1). https://doi.org/10.1186/s13058-015-0520-4
\item Fernandez-Martinez A, Krop IE, Hillman DW, Polley MY, Parker JS, Huebner L,
et al. \emph{Survival, pathologic response, and genomics in CALGB 40601
(Alliance), a neoadjuvant Phase III trial of paclitaxel-trastuzumab with or
without lapatinib in HER2-positive breast cancer.} Journal of Clinical
Oncology. 2020. https://doi.org/10.1200/JCO.20.01276
\item Zhao X, Rodland EA, Tibshirani R, Plevritis S. \emph{Molecular
subtyping for clinically defined breast cancer subgroups.} Breast Cancer
Research. 2015;17(1). https://doi.org/10.1186/s13058-015-0520-4
\item Fernandez-Martinez A, Krop IE, Hillman DW, Polley MY, Parker JS, Huebner L,
et al. \emph{Survival, pathologic response, and genomics in CALGB 40601
(Alliance), a neoadjuvant Phase III trial of paclitaxel-trastuzumab with or
without lapatinib in HER2-positive breast cancer.} Journal of Clinical
Oncology. 2020. https://doi.org/10.1200/JCO.20.01276
\item Paquet ER, Hallett MT. \emph{Absolute assignment of breast cancer
intrinsic molecular subtype.} J Natl Cancer Inst. 2015;107(1).
https://doi.org/10.1093/jnci/dju357
\item Staaf J, Häkkinen J, Hegardt C, Saal LH, Kimbung S, Hedenfalk I,
et al. \emph{RNA sequencing-based single sample predictors of molecular subtype
and risk of recurrence for clinical assessment of early-stage breast cancer}.
NPJ Breast Cancer. 2022;8(1). https://doi.org/10.1038/s41523-022-00465-3
}
}
