% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_Multi}
\alias{BS_Multi}
\title{BS_Multi}
\usage{
BS_Multi(
  data_input,
  phenodata,
  methods = NA,
  Subtype = FALSE,
  hasClinical = FALSE,
  ...
)
}
\arguments{
\item{data_input}{The output of Mapping() function.}

\item{phenodata}{A clinical information table. The first column must be named "PatientID".}

\item{methods}{Specify methods.}

\item{Subtype}{Logic. Specify whether to predict four subtypes by removing the Normal-like subtype.}

\item{hasClinical}{Logic. Specify whether clinical information is included. For example, tumor size should be in the "T" column, and lymph node status should be in the "NODE" column.}
}
\value{
The subtypes estimated by selected methods
}
\description{
This calls consensus subtyping and users can specify subtyping methods.
}
\examples{

data("OSLO2MEITOobj")
methods = c( "parker.median", "parker.mean", "parker.quantile")
res.test = BS_Multi(data_input = data_input, phenodata = clinic.oslo, methods = methods, Subtype = FALSE, hasClinical = FALSE)


}
