% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_Multi}
\alias{BS_Multi}
\title{Consensus Intrinsic Subtyping with Multiple Methods (BS_Multi)}
\usage{
BS_Multi(
  data_input,
  phenodata,
  methods = NA,
  Subtype = FALSE,
  hasClinical = FALSE,
  ...
)
}
\arguments{
\item{data_input}{The output from the \code{Mapping()} function, containing processed gene expression data prepared for subtyping analysis.}

\item{phenodata}{A clinical information table. The first column must be named "PatientID".}

\item{methods}{A character vector specifying the subtyping methods to be used. Options include:
\itemize{
\item "parker.original"
\item "genefu.scale"
\item "genefu.robust"
\item "ssBC"
\item "ssBC.v2"
\item "cIHC"
\item "cIHC.itr"
\item "PCAPAM50"
\item "AIMS"
\item "sspbc"
\item "AUTO" (automatically selects methods based on the biomarker distribution of the test cohort).
}

If "AUTO" is selected, it must be the only value in the vector. Otherwise, specify at least two other methods to perform subtyping. An error will be thrown if fewer than two methods (other than "AUTO") are provided.}

\item{Subtype}{Logical. If \code{TRUE}, predicts four subtypes by excluding the Normal-like subtype.}

\item{hasClinical}{Logical. If \code{TRUE}, includes clinical information in the analysis (e.g., tumor size in the "T" column and lymph node status in the "NODE" column).}
}
\value{
A list of intrinsic subtypes estimated by the selected methods.
}
\description{
Performs consensus intrinsic subtyping of breast cancer using multiple methods, allowing users to specify or automatically select subtyping approaches based on the test cohort's biomarker distribution.
}
\examples{
# Load required dataset
data("OSLO2EMIT0obj")

# Define methods to use for consensus subtyping
methods = c("parker.original", "genefu.scale", "genefu.robust")

# Perform subtyping
res.test = BS_Multi(
  data_input = data_input,
  phenodata = clinic.oslo,
  methods = methods,
  Subtype = FALSE,
  hasClinical = FALSE
)

}
