% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_parker}
\alias{BS_parker}
\title{BS_parker}
\usage{
BS_parker(
  gene_expression_matrix,
  phenodata = NA,
  calibration = "None",
  internal = NA,
  external = NA,
  medians = NA,
  Prosigna = FALSE,
  hasClinical = FALSE,
  ...
)
}
\arguments{
\item{gene_expression_matrix}{A gene expression matrix with genes in rows and samples in columns. The data should be log-transformed.}

\item{phenodata}{A clinical information table. The first column must be named "PatientID".}

\item{calibration}{The calibration method to use. Options are "None", "Internal", or "External". If "Internal" is selected, see the "internal" parameter for further details. If "External" is selected, see the "external" parameter.}

\item{internal}{Specify the strategy for internal calibration. Options are median-centered ("medianCtr", default), mean-centered ("meanCtr"), or quantile-centered ("qCtr").}

\item{external}{Specify the platform name (i.e., the column name) for external medians, which are calculated by the training cohort. If you want to use user-provided medians, set this parameter to "Given.mdns" and provide the medians via the "medians" parameter.}

\item{medians}{If "Given.mdns" is specified for the "external" parameter, input a matrix/table where the first column contains 50 genes and the second column contains the corresponding "Given.mdns" values.}

\item{Prosigna}{Logic.}

\item{hasClinical}{Logic. Specify whether clinical information is included. For example, tumor size should be in the "T" column, and lymph node status should be in the "NODE" column.}
}
\value{
The intrinsic subtypes estimated using the Parker-based method.
}
\description{
This calls parker-based intrinsic subtyping methods.
}
\examples{

data("oslo.obj")
res = BS_parker(data_input$x_NC.log,phenodata = NA, calibration = "Internal", internal = "medianCtr",  Prosigna = FALSE, hasClinical =FALSE)

}
