% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_parker}
\alias{BS_parker}
\title{Original Parker Intrinsic Subtyping (BS_parker)}
\usage{
BS_parker(
  se_obj,
  calibration = "None",
  internal = NA,
  external = NA,
  medians = NA,
  Subtype = FALSE,
  hasClinical = FALSE
)
}
\arguments{
\item{se_obj}{A \code{SummarizedExperiment} object containing:
\itemize{
\item \strong{Assay data}: A log-transformed, normalized gene expression matrix with genes (Gene Symbols) as rows
and samples as columns.
\item \strong{Column metadata} (\code{colData}): Optional clinical information.
}}

\item{calibration}{Specifies the calibration method to apply. Options include:
\itemize{
\item \code{"None"}: No calibration is applied.
\item \code{"Internal"}: Uses internal calibration strategies (see \code{internal} argument).
\item \code{"External"}: Uses external medians (see \code{external} argument).
}}

\item{internal}{Specifies the internal calibration strategy when \code{calibration = "Internal"}. Options include:
\itemize{
\item \code{"medianCtr"} (default): Median-centered calibration.
\item \code{"meanCtr"}: Mean-centered calibration (aligned with \code{genefu.scale}).
\item \code{"qCtr"}: Quantile-based calibration (aligned with \code{genefu.robust}).
}}

\item{external}{Specifies the platform name (i.e., column name) for external
medians derived from the training cohort.
\itemize{
\item To use user-provided medians, set \verb{"external = "Given.mdns"} and provide values via the
\code{medians} argument.
}}

\item{medians}{A matrix or table of user-provided median values, required if
\code{external = "Given.mdns"}.
\itemize{
\item The first column should contain 50 genes.
\item The second column should contain the corresponding median values.
}}

\item{Subtype}{Logical (\code{TRUE} or \code{FALSE}). If \code{TRUE}, the function predicts four subtypes,
\strong{excluding} the Normal-like subtype.}

\item{hasClinical}{Logical (\code{TRUE} or \code{FALSE}). If \code{TRUE}, the function incorporates clinical data from
the phenotype (\code{pheno}) table. Required columns:
\itemize{
\item \code{"TSIZE"}: Tumor size (\code{0} for <= 2cm, \code{1} for > 2cm).
\item \code{"NODE"}: Lymph node status (\code{0} for negative, \code{1} or higher for positive nodes; this column must be numeric).
}}
}
\value{
Returns a list containing intrinsic subtypes assigned using the Parker-based method, or its variations.
}
\description{
This function predicts breast cancer intrinsic subtypes using
the original Parker et al. (2019) method, along with variations of the original approach.
}
\examples{
data("OSLO2EMIT0obj")
res <- BS_parker(
    se_obj = OSLO2EMIT0obj$data_input$se_NC,
    calibration = "Internal",
    internal = "medianCtr",
    Subtype = FALSE,
    hasClinical = FALSE
)

}
\references{
\itemize{
\item Parker JS, Mullins M, Cheung MCU, Leung S, Voduc D, et al.
\emph{Supervised risk predictor of breast cancer based on intrinsic subtypes}.
Journal of Clinical Oncology. 2009;27(8).
https://doi.org/10.1200/JCO.2008.18.1370
\item Gendoo DMA, Ratanasirigulchai N, Schröder MS, Paré L, Parker JS, Prat A, et
al. \emph{Genefu: An R/Bioconductor package for computation of gene
expression-based signatures in breast cancer}. Bioinformatics. 2016;32(7).
https://doi.org/10.1093/bioinformatics/btv693
}
}
