% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_PCAPAM50}
\alias{BS_PCAPAM50}
\title{PCA-PAM50 Intrinsic Subtyping (BS_PCAPAM50)}
\usage{
BS_PCAPAM50(se_obj, Subtype = FALSE, hasClinical = FALSE, seed = 118)
}
\arguments{
\item{se_obj}{A \code{SummarizedExperiment} object containing:
\itemize{
\item \strong{Assay data}: A gene expression matrix with genes (Gene Symbol)
in rows and samples in columns. This matrix should be log-transformed.
\item \strong{Column metadata} (\code{colData}): Clinical information table.
\itemize{
\item \code{"PatientID"}: Unique sample or patient identifiers.
\item \code{"ER"}: Estrogen receptor (ER) status recorded as \code{"ER+"} or \code{"ER-"}.
}
}}

\item{Subtype}{Logical. If \code{TRUE}, the function predicts four subtypes by
excluding the Normal-like subtype.}

\item{hasClinical}{Logical. If \code{TRUE}, the function uses clinical data from
the \code{pheno} table. Required columns include:
\itemize{
\item \code{"TSIZE"}: Tumor size.
\item \code{"NODE"}: Lymph node status.
}}

\item{seed}{An integer value is used to set the random seed.}
}
\value{
A vector of intrinsic subtypes assigned to the samples, as estimated
by the PCA-PAM50 method.
}
\description{
This function predicts breast cancer intrinsic subtypes using
the PCA-PAM50 method. This approach integrates Principal Component Analysis
(PCA) to enhance the consistency of PAM50-based subtyping.
}
\examples{
data("OSLO2EMIT0obj")
res <- BS_PCAPAM50(
    se_obj = OSLO2EMIT0obj$data_input$se_NC,
    Subtype = FALSE,
    hasClinical = FALSE
)

}
\references{
\itemize{
\item Raj-Kumar PK, Liu J, Hooke JA, Kovatich AJ, Kvecher L, Shriver
CD, et al. \emph{PCA-PAM50 improves consistency between breast cancer intrinsic
and clinical subtyping, reclassifying a subset of luminal A tumors as luminal
B.} Sci Rep. 2019;9(1). https://doi.org/10.1038/s41598-019-44339-4
}
}
