% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_PCAPAM50}
\alias{BS_PCAPAM50}
\title{BS_PCAPAM50}
\usage{
BS_PCAPAM50(
  gene_expression_matrix,
  phenodata,
  Subtype = FALSE,
  hasClinical = FALSE,
  seed = 118
)
}
\arguments{
\item{gene_expression_matrix}{A gene expression matrix with genes in rows and samples in columns. The data should be log-transformed.}

\item{phenodata}{A clinical information table. The first column must be named "PatientID".}

\item{Subtype}{Logic. Specify whether to predict four subtypes by removing the Normal-like subtype.}

\item{hasClinical}{Logic. Specify whether clinical information is included. For example, tumor size should be in the "T" column, and lymph node status should be in the "NODE" column.}
}
\value{
The intrinsic subtypes estimated by PCA-PAM50 method.
}
\description{
This calls PCA-PAM50 to do intrinsic subtyping.
}
\examples{

data("OSLO2MEITOobj")
res = BS_PCAPAM50(data_input$x_NC.log, phenodata = clinic.oslo, Subtype = FALSE, hasClinical =FALSE, seed=118)

}
