% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_PCAPAM50}
\alias{BS_PCAPAM50}
\title{BS_PCAPAM50}
\usage{
BS_PCAPAM50(
  gene_expression_matrix,
  phenodata,
  Prosigna = FALSE,
  hasClinical = FALSE,
  seed = 118
)
}
\arguments{
\item{gene_expression_matrix}{Gene expression matrix, gene in row and sample in column}

\item{phenodata}{Clinical information table, with ER information or IHC column}

\item{hasClinical}{Logical. Please specify if you prepared clincical information, like Tumore size as T column, lymphatic node status as NODE column.}

\item{featuredata}{Annotate genes}
}
\value{
The intrinsic subtypes estimated by PCA-PAM50
}
\description{
This calls PCA-PAM50 (Raj-Kumar, PK.) to do intrinsic subtyping.
}
