% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_ssBC}
\alias{BS_ssBC}
\title{BS_ssBC}
\usage{
BS_ssBC(
  gene_expression_matrix,
  phenodata,
  s,
  Prosigna = FALSE,
  hasClinical = FALSE,
  ...
)
}
\arguments{
\item{gene_expression_matrix}{Gene expression matrix, gene in row and sample in column}

\item{phenodata}{Clinical information table, like TN,ER and HER2}

\item{s}{Specify "ER" or "TN" or "ER_JAMA" or "HER2+" or "TNBC". The original quantile is "ER" and "TN" of TNBC-BreastCancerRes2015.  If you choose "ER_JAMA" or "HER2+" or "TNBC", it means you choose quantile from TNBC-JAMAOncol2024.}

\item{hasClinical}{Logical. Please specify if you prepared clincical information, like Tumore size as T column, lymphatic node status as NODE column.}

\item{featuredata}{Annotate genes}
}
\value{
The subtypes estimated by ssBC
}
\description{
This calls ssBC to do intrinsic subtyping.
It is better to group cohorts by ER+/- or TN/nTN.
}
