% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_sspbc}
\alias{BS_sspbc}
\title{BS_sspbc}
\usage{
BS_sspbc(gene_expression_matrix, ...)
}
\arguments{
\item{gene_expression_matrix}{Gene expression matrix, gene in row sample in column}
}
\value{
The subtypes estimated by sspbc
}
\description{
This calls sspbc to do intrinsic subtyping.
}
