% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_sspbc}
\alias{BS_sspbc}
\title{BS_sspbc}
\usage{
BS_sspbc(gene_expression_matrix, ssp.name = "ssp.pam50", ...)
}
\arguments{
\item{gene_expression_matrix}{A gene expression matrix with genes in rows and samples in columns. The data should be log-transformed.}

\item{ssp.name}{Specify model names. Option is either "ssp.pam50" or "ssp.subtype". The latter one was used to predict four subtypes by removing the Normal-like subtype.}
}
\value{
The subtypes estimated by sspbc method
}
\description{
This calls sspbc to do intrinsic subtyping.
}
\examples{

data("OSLO2MEITOobj")
res = BS_sspbc( gene_expression_matrix = as.matrix(data_input$x_SSP), ssp.name= "ssp.pam50" )

}
\references{
Staaf J, H"kkinen J, Hegardt C, Saal LH, Kimbung S, Hedenfalk I, et al. RNA sequencing-based single sample predictors of molecular subtype and risk of recurrence for clinical assessment of early-stage breast cancer. NPJ Breast Cancer. 2022;8(1). https://doi.org/10.1038/s41523-022-00465-3
}
