% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{BS_sspbc}
\alias{BS_sspbc}
\title{Intrinsic Subtyping using SSPBC (BS_sspbc)}
\usage{
BS_sspbc(se_obj, ssp.name = "ssp.pam50")
}
\arguments{
\item{se_obj}{A \code{SummarizedExperiment} object containing:
\itemize{
\item \strong{Assay data}: A gene expression matrix with genes (EntrezID) and samples (columns).
Note: The data must be \strong{unlogged} before using this function.
}}

\item{ssp.name}{Specifies the model to use. Options are "ssp.pam50" (for
PAM50-based predictions) or "ssp.subtype" (for predicting four subtypes by
excluding the Normal-like subtype).}
}
\value{
A vector of intrinsic subtypes assigned to the samples, as estimated
by the SSPBC method.
}
\description{
This function predicts breast cancer intrinsic subtypes using
the SSPBC (Single Sample Predictor for Breast Cancer) method. This method is
based on a refined version of AIMS' original methodology, using a large,
uniformly accrued population-based cohort (SCAN-B) for training. Performs
intrinsic subtyping of breast cancer using the SSPBC (Single Sample Predictor
for Breast Cancer) method. This method supports predicting subtypes based on
RNA sequencing data and offers flexibility in selecting the prediction model.
}
\examples{

# Load required dataset
data("OSLO2EMIT0obj")

# Perform subtyping with the SSPBC method
res <- BS_sspbc(
    se_obj = OSLO2EMIT0obj$data_input$se_SSP,
    ssp.name = "ssp.pam50"
)

}
\references{
\itemize{
\item Staaf J, Häkkinen J, Hegardt C, Saal LH, Kimbung S, Hedenfalk I,
et al. \emph{RNA sequencing-based single sample predictors of molecular subtype and
risk of recurrence for clinical assessment of early-stage breast cancer}. NPJ
Breast Cancer. 2022;8(1). https://doi.org/10.1038/s41523-022-00465-3
}
}
