% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyBreastSubtypeR.R
\name{iBreastSubtypeR}
\alias{iBreastSubtypeR}
\title{iBreastSubtypeR}
\usage{
iBreastSubtypeR()
}
\value{
A table with subtyping and ROR score
}
\description{
Starts an interactive BreastSubtypeR shiny web app.

BreastSubtypeR integrates intrinsic molecular subtyping methods for breast cancer,
including nearest-centroid (NC-based) and single-sample predictor (SSP-based) approaches.
It employs standardized input and output formats,
offering a unified framework that is highly compatible with other R packages in the gene expression profiling field.

The \code{iBreastSubtypeR()} function starts an interactive shiny web app that allows
the user to configure the arguments of subtyping functions and
runs it on the local computer. Please see the manual page of the functions for a description of the arguments and their
default and alternative values.

Step 1:

The input data may be loaded from the users workspace or by selecting a CSV/text
file for the expression data, a CSV/text file for clinical information and
a CSV/text file for feature annotations.

After loading necessary files, users only need to run/click "Mapping" button once and waiting for the notification.
If Mapping() function runs well, it will say "You can go for Step 2.".

Step 2:

Then, users can select any subtyping method and change relevant parameters to perform analysis.
You will see "Analysis complete." when the analysis is finished.
Two plots would be visualized as well. Users can download the result as text file.
If you want to continue the analysis, you can choose another method directly and relevant parameters.
You don't have to run Step 1 again. You will have results of another method.
}
\examples{

iBreastSubtypeR() ## this will open your browser with the BreastSubtypeR shiny web app

}
\keyword{BreastSubtypeR}
\keyword{shiny}
