% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyBreastSubtypeR.R
\name{iBreastSubtypeR}
\alias{iBreastSubtypeR}
\title{iBreastSubtypeR}
\usage{
iBreastSubtypeR()
}
\value{
A table with subtyping and ROR score
}
\description{
Starts an interactive BreastSubtypeR shiny web app.

BreastSubtypeR collected breast cancer subtyping methods,
near centrioids based (NC-based) and single sample prediction based (SSP-based).
It employs standardized input and output and provides a unified framework that
is highly compatible with other R packages in the gene expression profiling field.

The \code{iBreastSubtypeR()} function starts an interactive shiny web app that allows
the user to configure the arguments of subtyping functions and
runs it on the computer. Please see the manual page of the functions for a description of the arguments and their
default and alternative values.

The input data may be loaded from the users workspace or by selecting a CSV/text
file for the expression data, a CSV/text file for clinical information and
a CSV/text file for feature annotations.
}
\examples{
\dontrun{
res <- iBreastSubtypeR() ## this will open your browser with the GSVA shiny web app
}

}
\keyword{BreastSubtypeR}
\keyword{shiny}
