% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{Mapping}
\alias{Mapping}
\title{Mapping}
\usage{
Mapping(
  gene_expression_matrix,
  featuredata,
  method = "mean",
  mapping = TRUE,
  impute = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{gene_expression_matrix}{Gene expression matrix. Probe or gene symbol in row and sample in column.}

\item{featuredata}{Feature data provided by user. The table should contain at least three column, which are probe(probeid or transcriptID), EntrezGene.ID and symbol.}

\item{method}{Method to deduplicated probes for microarray or RNAseq. Please select "IQR" for Affy, "mean" for Agilent and "max" for RNAseq.}

\item{impute}{Logic. Please specify whether to perform impute.knn on NA data}

\item{verbose}{Logic.}
}
\description{
function for mapping gene ID
}
