% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{Mapping}
\alias{Mapping}
\title{Gene ID Mapping}
\usage{
Mapping(
  gene_expr,
  featuredata = NA,
  method = "max",
  impute = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gene_expr}{Gene expression matrix with probes in rows and
samples in columns.}

\item{featuredata}{Feature data provided by the user. The table must contain
at least three columns: \code{probe} (ProbeID or TranscriptID), \code{SYMBOL}, and
\code{ENTREZID}.}

\item{method}{Method for deduplicating probes in microarray or RNA-seq data.
Choose one of the following options:
\itemize{
\item \code{"IQR"} for Affymetrix arrays,
\item \code{"mean"} for Agilent/Illumina arrays,
\item \code{"max"} for RNA-seq data.
}}

\item{impute}{Logical. Specify whether to perform K-Nearest Neighbors (KNN)
imputation on missing data (\code{NA} values).}

\item{verbose}{Logical. If \code{TRUE}, progress messages will be displayed during
execution.}
}
\value{
A list containing three gene expression matrices:
\itemize{
\item \code{"x_NC.log"}:
Preprocessed matrix for input into nearest-centroid (NC)-based methods.
\item \code{"x_SSP"}: Preprocessed matrix for input into single-sample predictor
(SSP)-based methods.
}
}
\description{
A function to map gene IDs and preprocess gene expression data
for subsequent analyses.
}
\details{
If the rows of the gene expression matrix/table represent gene
symbols, an additional \code{SYMBOL} column must be added to the feature table and
renamed as \code{probe}.
}
\examples{
data("OSLO2EMIT0obj")
data_input <- Mapping(
    gene_expr = OSLO2EMIT0obj$OSLO2EMIT0.103.genematrix_noNeg.subset,
    featuredata = OSLO2EMIT0obj$anno_feature.subset,
    method = "max",
    impute = TRUE,
    verbose = TRUE
)

}
