% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R, R/Utilities.R
\name{Mapping}
\alias{Mapping}
\title{Mapping}
\usage{
Mapping(
  gene_expression_matrix,
  featuredata = NA,
  method = "max",
  impute = TRUE,
  verbose = TRUE
)

Mapping(
  gene_expression_matrix,
  featuredata = NA,
  method = "max",
  impute = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gene_expression_matrix}{Gene expression matrix}

\item{featuredata}{Feature data provided by user. The table should contain at least two column, probe and ENTREZID.}

\item{method}{Method to deduplicated probes for microarray or RNAseq. Please select "IQR" for Affy, "mean" for Agilent and max for RNAseq}

\item{impute}{Logic. Please specify if there are NA data adn want keep them}

\item{verbose}{Logic.}
}
\value{
lists of three gene expression matrix. "x_NC.log" is prepared for NC-based methods and "x_SSP" for SSP-based methods.
}
\description{
function for mapping gene ID
}
\details{
The row can be gene symbol names in gene expression matrix/table, but you need to add one extra SYMBOL column and rename it as probe in feature table.
}
\examples{

data("OSLO2MEITOobj")
data = OSLO2EMIT0.103.genematrix_noNeg[,clinic.oslo$PatientID]
data_input = Mapping(gene_expression_matrix = data, featuredata = anno_feature, impute = TRUE, verbose = TRUE )

}
