% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PAM50_subtyping_methods.R
\name{Mapping}
\alias{Mapping}
\title{Mapping}
\usage{
Mapping(
  x,
  y,
  method = "mean",
  mapping = TRUE,
  impute = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{x}{Gene expression matrix}

\item{y}{Feature data provided by user. The table should contain at least three column, which are probe(probeid or transcriptID), EntrezGene.ID and symbol.}

\item{method}{Method to deduplicated probes for microarray or RNAseq. Please select IQR for Affy and select mean for Agilent)}

\item{impute}{Logic. Please specify if there are NA data adn want keep them}

\item{verbose}{Logic.}
}
\description{
function for mapping ID and supplementing missing data if necessary
}
