% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R, R/Utilities.R
\name{Mapping}
\alias{Mapping}
\title{Mapping}
\usage{
Mapping(
  gene_expression_matrix,
  featuredata = NA,
  method = "max",
  impute = TRUE,
  verbose = TRUE
)

Mapping(
  gene_expression_matrix,
  featuredata = NA,
  method = "max",
  impute = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gene_expression_matrix}{Gene expression matrix}

\item{featuredata}{Feature data provided by user. The table should contain at least two column, probe and ENTREZID.}

\item{method}{Method to deduplicated probes for microarray or RNAseq. Please select "IQR" for Affy, "mean" for Agilent and max for RNAseq}

\item{impute}{Logic. Please specify if there are NA data adn want keep them}

\item{verbose}{Logic.}
}
\description{
function for mapping gene ID
}
