% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R
\name{Mapping}
\alias{Mapping}
\title{Gene ID Mapping}
\usage{
Mapping(se_obj, method = "max", impute = TRUE, verbose = TRUE)
}
\arguments{
\item{se_obj}{A \code{SummarizedExperiment} object containing:
\itemize{
\item \strong{Assay data}: A gene expression matrix where rows represent probes (e.g., ProbeID, TranscriptID, or Gene Symbol)
and columns represent samples. This should be stored in the \code{assay()} slot.
\item \strong{Row metadata}: A data frame with annotations for probes, including at least the following columns:
\itemize{
\item \code{"probe"}: Unique identifiers for the probes (e.g., ProbeID, TranscriptID or gene symbol).
\item \code{"ENTREZID"}: Entrez gene IDs corresponding to the probes.
}
\item \strong{Column metadata} (optional): Sample metadata stored in the \code{colData()} slot.
}}

\item{method}{Method for deduplicating probes in microarray or RNA-seq data.
Choose one of the following options:
\itemize{
\item \code{"IQR"} for Affymetrix arrays,
\item \code{"mean"} for Agilent/Illumina arrays,
\item \code{"max"} for RNA-seq data.
}}

\item{impute}{Logical. Specify whether to perform K-Nearest Neighbors (KNN)
imputation on missing data (\code{NA} values).}

\item{verbose}{Logical. If \code{TRUE}, progress messages will be displayed during
execution.}
}
\value{
A list containing three preprocessed gene expression matrices:
\itemize{
\item \code{"x_NC"}: The processed gene expression matrix
for nearest-centroid (NC)-based methods.
\item \code{"x_NC.log"}: The log-transformed version
of \code{"x_NC"} for NC-based methods.
\item \code{"x_SSP"}: The processed gene expression matrix
for single-sample predictor (SSP)-based methods.
}
}
\description{
A function to map gene IDs and preprocess gene expression data
for subsequent analyses.
}
\details{
If the rows of the gene expression matrix/table represent gene
symbols, an additional \code{SYMBOL} column must be added to the feature table and
renamed as \code{probe}.
}
\examples{
data("OSLO2EMIT0obj")
data_input <- Mapping(
    se_obj = OSLO2EMIT0obj$se_obj,
    method = "max",
    impute = TRUE,
    verbose = TRUE
)

}
