% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BreastSubtypeR_methods.R, R/Utilities.R
\name{Mapping}
\alias{Mapping}
\title{Gene ID Mapping}
\usage{
Mapping(
  gene_expression_matrix,
  featuredata = NA,
  method = "max",
  impute = TRUE,
  verbose = TRUE
)

Mapping(
  gene_expression_matrix,
  featuredata = NA,
  method = "max",
  impute = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{gene_expression_matrix}{A matrix of gene expression data with genes in rows and samples in columns. The data should be pre-processed and log-transformed.}

\item{featuredata}{A table containing feature data with at least two columns:
\itemize{
\item \code{"probe"}: The probe identifiers (e.g., probe names).
\item \code{"ENTREZID"}: Corresponding ENTREZ gene IDs.
}}

\item{method}{A character string specifying the method to deduplicate probes:
\itemize{
\item \code{"IQR"}: Uses interquartile range (IQR) to select a representative probe (preferred for short oligo, e.g., for Affymetrix microarrays).
\item \code{"mean"}: Uses the mean expression of probes that map to the same gene (preferred for long oligo, e.g., for Agilent/Illumina microarrays).
\item \code{"max"}: Selects the probe with the highest expression (e.g., for RNA-seq).
}}

\item{impute}{Logical. If \code{TRUE}, missing values (NA) in the gene expression matrix are handled (e.g., imputation may occur). If \code{FALSE}, rows with NA values are excluded.}

\item{verbose}{Logical. If \code{TRUE}, detailed information on the process is printed during execution.}
}
\value{
A list containing three gene expression matrices:
\itemize{
\item \code{"x_NC.log"}: Preprocessed matrix for input into nearest-centroid (NC)-based methods.
\item \code{"x_SSP"}: Preprocessed matrix for input into single-sample predictor (SSP)-based methods.
}

A gene expression matrix with mapped ENTREZ IDs, and optionally imputed missing values. The matrix will have genes in rows and samples in columns, with ENTREZ gene IDs as row names.
}
\description{
A function to map gene IDs and preprocess gene expression data for subsequent analyses.

This function maps gene IDs from the provided gene expression matrix to ENTREZ IDs and handles missing data if necessary. It includes options for deduplicating probes based on the selected method.
}
\details{
If the rows of the gene expression matrix/table represent gene symbols, an additional \code{SYMBOL} column must be added to the feature table and renamed as \code{probe}.
}
\examples{
data("OSLO2EMIT0obj")
data = OSLO2EMITO.103.genematrix_noNeg.subset[, clinic.oslo$PatientID]
data_input = Mapping(
  gene_expression_matrix = data,
  featuredata = anno_feature,
  method = "max",
  impute = TRUE,
  verbose = TRUE
)

data("gene_expression_data")
data("feature_data")
res <- MapGeneIDs(
  gene_expression_matrix = gene_expression_data,
  featuredata = feature_data,
  method = "IQR",
  impute = TRUE,
  verbose = TRUE
)

}
