% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{sspbc.models}
\alias{sspbc.models}
\title{The sspbc models (short name) for the 11 developed predictors}
\format{
An object of class \code{list} of length 11.
}
\usage{
data("sspbc.models")
}
\value{
\item{sspbc.models}{The collection of 11 ssp models used by sspbc.}

#' @references
\itemize{
\item Staaf J, Häkkinen J, Hegardt C, Saal LH, Kimbung S, Hedenfalk I,
et al. \emph{RNA sequencing-based single sample predictors of molecular subtype and
risk of recurrence for clinical assessment of early-stage breast cancer}. NPJ
Breast Cancer. 2022;8(1). https://doi.org/10.1038/s41523-022-00465-3
}
}
\description{
List with the 11 SSP models from Staaf J. et al. medRxiv 2021.12.03.21267116.
}
\details{
List elements are named with short name for respective ssp model.
The ssp models in list sspbc.models are the same as in list sspbc.models.fullname.
}
\examples{
library(BreastSubtypeR)
## Load the sspbc.models
data("sspbc.models")

}
\keyword{datasets}
