% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC_functions.R
\name{Vis_boxpot}
\alias{Vis_boxpot}
\title{Functions for visualization
Function for boxplot of correlation per subtype}
\usage{
Vis_boxpot(out, correlations)
}
\arguments{
\item{out}{a data frame includes "patientID" and "Subtype"}

\item{correlations}{correlations table from NC-based methods}
}
\description{
Functions for visualization
Function for boxplot of correlation per subtype
}
