% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Vis_consensus}
\alias{Vis_consensus}
\title{Function for pieplot}
\usage{
Vis_consensus(data)
}
\arguments{
\item{out}{the output of BS_Check() function.}
}
\description{
This function is used to visualize the intrinsic subtypes estimated by BS_Check() function.
}
\examples{

data("OSLO2MEITOobj")
plot = Vis_consensus(data = res$res_subtypes)

}
