% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NC_functions.R
\name{Vis_heatmap}
\alias{Vis_heatmap}
\title{Function for heatmap visualizayion}
\usage{
Vis_heatmap(x, out)
}
\arguments{
\item{x}{gene expression matrix, log2 transformed}

\item{out}{a data frame includes "patientID" and "Subtype"}
}
\description{
Function for heatmap visualizayion
}
