% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Vis_heatmap}
\alias{Vis_heatmap}
\title{Function for heatmap visualizayion}
\usage{
Vis_heatmap(x, out)
}
\arguments{
\item{x}{gene expression matrix, log2 transformed}

\item{out}{a data frame includes "patientID" and "Subtype"}
}
\description{
Function for heatmap visualizayion
}
\examples{

data("OSLO2MEITOobj")

x = data_input$x_NC
out= data.frame(PatientID = res$results$parker.original$BS.all$PatientID, Subtype = res$results$parker.original$BS.all$BS )

p = Vis_heatmap(x, out)
plot(p)

}
