% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Vis_Multi}
\alias{Vis_Multi}
\title{Function for pieplot}
\usage{
Vis_Multi(data)
}
\arguments{
\item{out}{a data table includes "patientID" and "Subtype"}
}
\description{
Function for pieplot
}
\examples{

data("OSLO2MEITOobj")
p = Vis_Multi(res$res_subtypes)
plot(p)
}
