% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utilities.R
\name{Vis_PCA}
\alias{Vis_PCA}
\title{Function for PCAplot}
\usage{
Vis_PCA(x, out, Eigen = FALSE)
}
\arguments{
\item{x}{gene expression matrix, log2 transformed}

\item{out}{a data table includes "patientID" and "Subtype"}

\item{Eigen}{Logic. Please specify if show screeplot}
}
\description{
Function for PCAplot
}
\examples{


data("OSLO2MEITOobj")

x = data_input$x_NC.log
out = data.frame(PatientID = res$results$parker.original$BS.all$PatientID, Subtype = res$results$parker.original$BS.all$BS )
p = Vis_PCA(x = x, out = out)
plot(p)

}
